requirements.txt
nuscenes-devkit
torch 
torchvision 
torchaudio
matplotlib
seaborn
pillow
numpy
segmentation-models-pytorch
ultralytics
ipython